src/test/java/com/rtg/ngs/LongReadMaskParamsTest.java
/*
 * Copyright (c) 2018. Real Time Genomics Limited.
 *
 * All rights reserved.
 *
 * Redistribution and use in source and binary forms, with or without
 * modification, are permitted provided that the following conditions are
 * met:
 *
 * 1. Redistributions of source code must retain the above copyright
 *    notice, this list of conditions and the following disclaimer.
 *
 * 2. Redistributions in binary form must reproduce the above copyright
 *    notice, this list of conditions and the following disclaimer in the
 *    documentation and/or other materials provided with the
 *    distribution.
 *
 * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
 * "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
 * LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR
 * A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT
 * HOLDER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL,
 * SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT
 * LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
 * DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY
 * THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
 * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
 * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 */
package com.rtg.ngs;

import junit.framework.TestCase;

/**
 * Tests corresponding class
 */
public class LongReadMaskParamsTest extends TestCase {

  public void testGetters() {
    final int readLength = 35;
    final LongReadMaskParams lrmp = new LongReadMaskParams(25, 1, 2, 3);
    assertEquals(25, lrmp.getWordSize());
    assertEquals(1, lrmp.getSubstitutions());
    assertEquals(2, lrmp.getIndels());
    assertEquals(3, lrmp.getIndelLength());

    assertTrue(lrmp.isValid(readLength));
    assertEquals("Long read Mask: w=25 s=1 i=2 l=3", lrmp.toString());

    try {
      lrmp.maskFactory(readLength);
      fail();
    } catch (UnsupportedOperationException ex) {
      assertEquals("Not supported, ever", ex.getMessage());
    }
  }


}
